+++ app/Filament/Pages/NoScrum.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Pages;
+
+use Filament\Pages\Page;
+
+class NoScrum extends Page
+{
+    public static bool $shouldRegisterNavigation = false;
+    protected static string $view = 'filament.pages.no-scrum';
+}